+++ tests/lints/check-capi-docs.py
@@ -38,6 +38,7 @@ def functions_in_outline():
         "chfl_clear_errors",
         "chfl_version",
         "chfl_formats_list",
+        "chfl_guess_format",
     ]
     DOCS = os.path.join(ROOT, "doc", "src", "capi")
     functions = MISC_FUNCTIONS